[kernel] Build everything with -Wconversion [part 6]
Changes to be committed:
	modified:   common/arch/generic_x86/cpu_features.c
	modified:   include/tilck/kernel/kb.h
	modified:   kernel/arch/generic_x86/8042/kb.c
	modified:   kernel/arch/generic_x86/8042/kb_ansi_seq.c.h
	modified:   kernel/arch/i386/process32.c

+++ common/arch/generic_x86/cpu_features.c
@@ -148,7 +148,7 @@ void dump_x86_features(void)
             continue;
 
          if (flags[j][i])
-            w += snprintk(buf + w, sizeof(buf) - w, "%s ", strings[j][i]);
+            w += (u32)snprintk(buf + w, sizeof(buf) - w, "%s ", strings[j][i]);
 
          if (w >= 60) {
             printk("%s\n", buf);
@@ -159,7 +159,7 @@ void dump_x86_features(void)
    }
 
    if (x86_cpu_features.avx2)
-      w += snprintk(buf + w, sizeof(buf) - w, "avx2 ");
+      w += (u32)snprintk(buf + w, sizeof(buf) - w, "avx2 ");
 
    if (w)
       printk("%s\n", buf);
