+++ include/tilck/kernel/kb.h
@@ -14,7 +14,7 @@ typedef int (*keypress_func)(u32, u8);
 void init_kb(void);
 bool kb_is_pressed(u32 key);
 int kb_register_keypress_handler(keypress_func f);
-bool kb_scancode_to_ansi_seq(u32 key, u32 modifiers, char *seq);
+bool kb_scancode_to_ansi_seq(u32 key, u8 modifiers, char *seq);
 u32 kb_get_current_modifiers(void);
 int kb_get_fn_key_pressed(u32 key);
